+++ globals/foraging.go
@@ -1,5 +1,10 @@
 package globals
 
+type ForagingPeriodData struct {
+	Years       [][]float64
+	CurrentYear int
+}
+
 type ForagingPeriod struct {
 	SecondsToday int
 }