+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Implements seasonal and "scripted" patch dynamics (#49)
 - Adds support for weather/foraging period from files (#50)
 - Adds a model initializer for using custom systems (#53)
+- Foraging period files can contain multiple years of data (#60)
 
 ### Bugfixes